finishing mylist

+++ lib/modules/my_list/components/my_list_button.dart
@@ -1,6 +1,9 @@
 import 'package:anilist/constant/app_color.dart';
 import 'package:anilist/constant/divider.dart';
+import 'package:anilist/core/routes/route.dart';
+import 'package:anilist/global/bloc/app_bloc/app_bloc.dart';
 import 'package:anilist/global/model/anime.dart';
+import 'package:anilist/modules/auth/screen/login_screen.dart';
 import 'package:anilist/modules/my_list/bloc/my_list_bloc.dart';
 import 'package:anilist/utils/view_utils.dart';
 import 'package:anilist/widget/text/text_widget.dart';
@@ -57,6 +60,18 @@ class _MyListButtonState extends State<MyListButton> {
           color: Colors.transparent,
           child: InkWell(
             onTap: () {
+              if (context.read<AppBloc>().state.user == null) {
+                showConfirmationDialog(
+                  context: context,
+                  title: 'Access Denied',
+                  description: 'Please log in to continue.',
+                  okText: 'Log In',
+                  onTapOk: () =>
+                      pushAndRemoveUntil(context, screen: LoginScreen()),
+                );
+                return;
+              }
+
               if (_isMyList) {
                 _myListBloc.add(DeleteMyListEvent(widget.anime.malId));
               } else {
